Sporo poprawek, wstep do systemu obsługi projektu: - todo: zapisywanie /wczytywanie projektu, ewentualne blokady, toolbar

+++ trunk/Program/main.cpp
@@ -10,6 +10,7 @@
 #include "splash.h"
 //---------------------------------------------------------------------------
 #pragma package(smart_init)
+#pragma link "cgauges"
 #pragma resource "*.dfm"
 TForm1 *Form1;
 //---------------------------------------------------------------------------
@@ -20,15 +21,21 @@ __fastcall TForm1::TForm1(TComponent* Owner)
 
 void TForm1::OnLoadProgress(void* Sender, int position, int max, AnsiString info, int id)
 {
-	if (id==0) Form2->log->Caption="Zakoñczono wczytywanie typu: "+info;
-	if (id==1) Form2->log->Caption="Zakoñczono wczytywanie funkcji: "+info;
+	if (id==0) {
+				 Form2->log->Caption="Zakoñczono wczytywanie typu: "+info;
+				 OnLogSuccess(NULL,"Wczytano typ: "+info);
+			   }
+	if (id==1) {
+				 Form2->log->Caption="Zakoñczono wczytywanie funkcji: "+info;
+				 OnLogSuccess(NULL,"Wczytano funkcjê: "+info);
+			   }
 	if (id==2) {
-				Form2->log->Caption="B³êdna DLL typu: "+info;
-				ShowMessage("B³êdna DLL typu: "+info);
+				 Form2->log->Caption="B³êdna DLL typu: "+info;
+				 OnLogWarrning(NULL,"B³¹d podczas wczytywania typu, b³êdna DLL: "+info);
 			   }
 	if (id==3) {
-				Form2->log->Caption="B³êdna DLL funkcji: "+info;
-				ShowMessage("B³êdna DLL funckji: "+info);
+				 Form2->log->Caption="B³êdna DLL funkcji: "+info;
+				 OnLogWarrning(NULL,"B³¹d podczas wczytywania funkcji, b³êdna DLL: "+info);
 			   }
 	Form2->ProgressBar1->Position=position;
 	Form2->ProgressBar1->Max=max+1;
@@ -46,36 +53,30 @@ void __fastcall TForm1::FormCreate(TObject *Sender)
 	Application->CreateForm(__classid(TForm2), &Form2);
 	//Form1->DoubleBuffered=true;
 	Form2->ProgressBar1->Visible=false;
-    Form2->Visible=true;
+	Form2->Visible=true;
+	piwo=NULL;
 	AnsiString dir=ExtractFileDir(Application->ExeName);
 	Form2->ProgressBar1->Visible=true;
+    OnLogInformation(NULL,"Rozpoczêto wczytywanie pluginów");
 	plugins.OnLoadingProgress=OnLoadProgress;
 	plugins.OnFunctionAddRequest=OnFunctionAddClick;
 	plugins.LoadData(dir+"\\blocks", dir+"\\types", *MainMenu1, *ImageList1, 3, 2, 0, 1);
+	OnLogInformation(NULL,"Zakoñczono wczytywanie pluginów");
+
 	Form2->log->Caption="Wczytywanie interfejsu";
 	Form2->ProgressBar1->Position=Form2->ProgressBar1->Max;
 	Application->ProcessMessages();
+	blockMenu(true);
 	//Tworzenie szablonu PIWOEngine;
-	piwo=new PIWOEngine(this);
-	piwo->plugins=&plugins;
-	piwo->Align=alClient;
-	piwo->Parent=this;
-	piwo->OnInformation=OnLogInformation;
-	piwo->OnWarrning=OnLogWarrning;
-	piwo->OnError=OnLogError;
-	piwo->OnDebug=OnLogDebug;
-	piwo->OnSuccess=OnLogSuccess;
-	piwo->OnRunInformation=OnLogRunInformation;
-	piwo->OnRunWarrning=OnLogRunWarrning;
-	piwo->OnRunError=OnLogRunError;
-	piwo->OnRunDebug=OnLogRunDebug;
-	piwo->OnRunSuccess=OnLogRunSuccess;
-    PageControl1Resize(NULL);
+	newProject();
+
 	Sleep(500);
 	Form1->Visible=true;
 	Form1->Enabled=false;
 	Application->ProcessMessages();
+	PageControl1Change(NULL);
 	Sleep(500);
+
 	Form2->Visible=false;
 	Form1->Enabled=true;
 	delete Form2;
@@ -353,7 +354,9 @@ void __fastcall TForm1::Zresetujzaznaczonepoczenie1Click(TObject *Sender)
 
 void __fastcall TForm1::Uruchomwszystko1Click(TObject *Sender)
 {
- piwo->run(false);
+  blockMenu(true);
+  piwo->run(false);
+  blockMenu(false);
 }
 //---------------------------------------------------------------------------
 
@@ -367,13 +370,13 @@ void __fastcall TForm1::Uruchom3Click(TObject *Sender)
 
 void __fastcall TForm1::PageControl1Resize(TObject *Sender)
 {
-   PageControl1->TabWidth=int((PageControl1->Width-30)/3);
+   PageControl1->TabWidth=int((PageControl1->Width-40)/3);
    ListView1->Width=TabSheet1->Width;
    ListView2->Width=TabSheet2->Width;
    ListView3->Width=TabSheet3->Width;
+   Label1->Width=int((PageControl1->Width-40)/3)+5;
 }
 //---------------------------------------------------------------------------
-
 
 void __fastcall TForm1::Wyczylogi1Click(TObject *Sender)
 {
@@ -474,7 +477,7 @@ void __fastcall TForm1::MenuItem4Click(TObject *Sender)
 void __fastcall TForm1::FormKeyDown(TObject *Sender, WORD &Key,
 	  TShiftState Shift)
 {
-	if (Key==VK_DELETE) {
+	if (Key==VK_DELETE&&piwo!=NULL&&piwo->Enabled) {
 	   piwo->DeleteSelectedConnection();
 	   piwo->DeleteSelectedBlocks(); 
 	}
@@ -490,6 +493,129 @@ void __fastcall TForm1::PageControl1Change(TObject *Sender)
   ListView2->Scroll(0,99999999);
   ListView1->Scroll(0,99999999);
   PageControl1Resize(NULL);
+}
+//---------------------------------------------------------------------------
+void TForm1::OnRunStart(TObject* Sender)
+{
+	Timer1->Enabled=true;
+	Label1->Caption="";
+	CGauge1->Progress=0;
+	Application->ProcessMessages();
+}
+
+void TForm1::OnRunEnd(TObject* Sender)
+{
+	Panel2->Visible=false;
+	Timer1->Enabled=false;
+	Uruchom3->Enabled=true;
+	Uruchomwszystko1->Enabled=true;
+	Anuluj1->Enabled=false;
+}
+
+void TForm1::OnRunProgress(TObject* Sender,const AnsiString message, const double precent)
+{
+	Label1->Caption=message;
+	CGauge1->Progress=int(precent);
+	Application->ProcessMessages();
+}
+
+void __fastcall TForm1::SpeedButton1Click(TObject *Sender)
+{
+	if (piwo!=NULL)
+	{
+       piwo->abort();
+	}
+}
+//---------------------------------------------------------------------------
+
+void __fastcall TForm1::Timer1Timer(TObject *Sender)
+{
+  Panel2->Visible=true;
+  Timer1->Enabled=false;
+}
+
+void TForm1::blockMenu(bool blocked)
+{
+  if (piwo!=NULL)
+  {
+	  plugins.setMenuItemsStatus(!blocked);
+	  Uruchom3->Enabled=!blocked;
+	  Uruchomwszystko1->Enabled=!blocked;
+	  Anuluj1->Enabled=blocked;
+	  Zaznaczwszystkiebloki1->Enabled=!blocked;
+	  Odznaczwszystkiebloki1->Enabled=!blocked;
+	  Odwrzaznaczenieblokw1->Enabled=!blocked;
+	  Duplikujbloki1->Enabled=!blocked;
+	  Usubloki1->Enabled=!blocked;
+	  Usuzaznaczonebloki1->Enabled=!blocked;
+	  Odznaczzaznaczonepoaczenie1->Enabled=!blocked;
+	  Usuwszystkiepoczenia1->Enabled=!blocked;
+	  Usuzaznaczonepoczenie1->Enabled=!blocked;
+	  Zresetujwszystkiepoczenia1->Enabled=!blocked;
+	  Zresetujzaznaczonepoczenie1->Enabled=!blocked;
+	  Zapiszjako1->Enabled=!blocked;
+	  Exportujjakoobraz1->Enabled=!blocked;
+	  Zakocz1->Enabled=!blocked;
+	  piwo->Enabled=!blocked;
+  }
+  else
+  {
+	  plugins.setMenuItemsStatus(false);
+	  Uruchom3->Enabled=false;
+	  Uruchomwszystko1->Enabled=false;
+	  Anuluj1->Enabled=false;
+	  Zaznaczwszystkiebloki1->Enabled=false;
+	  Odznaczwszystkiebloki1->Enabled=false;
+	  Odwrzaznaczenieblokw1->Enabled=false;
+	  Duplikujbloki1->Enabled=false;
+	  Usubloki1->Enabled=false;
+	  Usuzaznaczonebloki1->Enabled=false;
+	  Odznaczzaznaczonepoaczenie1->Enabled=false;
+	  Usuwszystkiepoczenia1->Enabled=false;
+	  Usuzaznaczonepoczenie1->Enabled=false;
+	  Zresetujwszystkiepoczenia1->Enabled=false;
+	  Zresetujzaznaczonepoczenie1->Enabled=false;
+	  Zapiszjako1->Enabled=false;
+	  Exportujjakoobraz1->Enabled=false;
+	  Zakocz1->Enabled=false;
+  }
+}
+
+void TForm1::newProject()
+{
+    if (!closeProject()) return;
+	fileName="";
+	piwo=new PIWOEngine(this);
+	piwo->plugins=&plugins;
+	piwo->Align=alClient;
+	piwo->Parent=this;
+	piwo->OnInformation=OnLogInformation;
+	piwo->OnWarrning=OnLogWarrning;
+	piwo->OnError=OnLogError;
+	piwo->OnDebug=OnLogDebug;
+	piwo->OnSuccess=OnLogSuccess;
+	piwo->OnRunInformation=OnLogRunInformation;
+	piwo->OnRunWarrning=OnLogRunWarrning;
+	piwo->OnRunError=OnLogRunError;
+	piwo->OnRunDebug=OnLogRunDebug;
+	piwo->OnRunSuccess=OnLogRunSuccess;
+	piwo->OnRunStart=OnRunStart;
+	piwo->OnRunEnd=OnRunEnd;
+	piwo->OnRunProgress=OnRunProgress;
+	blockMenu(false);
+	Caption=CAPTION;
+	Caption+=" - Nowy projekt";
+	StatusBar1->SimpleText="Nowy projekt";
+}
+
+bool TForm1::closeProject()
+{
+   return true;
+}
+
+void __fastcall TForm1::Nowy1Click(TObject *Sender)
+{
+	newProject();
 }
 //---------------------------------------------------------------------------
